scan fixes

+++ phase2-init/src/MyFileClient.java
@@ -7,8 +7,9 @@ public class MyFileClient
 	public boolean startMyFileClient(String server, int port, Token token)
 	{
 		boolean flag = true;
+		boolean connected = fileclient.connect(server, port);
 		while (flag){
-			if (fileclient.connect(server, port))
+			if (connected)
 			{
 				Scanner sc = new Scanner(System.in);
 			
@@ -21,20 +22,24 @@ public class MyFileClient
 				System.out.println("(5) - Disconnect from File Server");
 			
 				int commandNum = sc.nextInt();
-			
-				switch (commandNum) 
-				{
+				sc.nextLine();
+				
 				// list files
-				case 1: 
+				if (commandNum == 1) {
 					List<String> filelist;
 					filelist = fileclient.listFiles(token);
 					System.out.println("(1) File List");
-					for (String f:filelist){
-						System.out.println(f);
+					if (filelist != null){
+						for (String f:filelist){
+							System.out.println(f);
+						}
 					}
-					break;
+					else {
+						System.out.println("No files");
+					}
+				}
 				// upload file
-				case 2:  
+				else if (commandNum == 2) {
 					System.out.println("(2) Upload File");
 					System.out.println("Enter source file: ");
 					String sourcefile = sc.nextLine();
@@ -45,9 +50,9 @@ public class MyFileClient
 					System.out.println("Enter group that file will belong to: ");
 					String grp = sc.nextLine();
 					fileclient.upload(sourcefile, destfile, grp, token);
-					break;
+				}
 				// download file
-				case 3:
+				if (commandNum == 3) {
 					System.out.println("(3) Download File");
 					System.out.println("Enter source file: ");
 					String sFile = sc.nextLine();
@@ -56,22 +61,23 @@ public class MyFileClient
 					String dFile = sc.nextLine();
 
 					fileclient.download(sFile, dFile, token);
-					break;
+				}
 				// delete file
-				case 4:
+				if (commandNum == 4) {
 					System.out.println("(4) Delete File");
 					System.out.println("Enter name of file to be deleted: ");
 					String filename = sc.nextLine();
 					fileclient.delete(filename, token);
-					break;
+				}
 				// disconnect from server
-				case 5:
+				if (commandNum == 5) {
 					System.out.println("(5) Disconnecting From File Server");
 					fileclient.disconnect();
+					connected = false;
 					flag = false;
-					break;
-				default: System.out.println("Invalid Command Number");
-					break;
+				}
+				else {
+					System.out.println("Invalid Command Number");
 				}
 			
 			}
